Fix #77 Ensure slashes remain slashes in encoded object URL
Unlike directory queries, which are sent in the query part of the
request, the access to objects is sent in the path part. Thus, we
must not encode slashes as they are an essential component of the full
path to an object.

This patch does the full encoding of the path, and then reverts all
occurrences of %2F to a plain '/'.