Configure jackson ObjectMapper to work in secure environment
By default jackson requires supressAccessChecks permission to fix visilibity
on some members/constructors. This is not really needed in the context of
this project but can cause issues when joss is used with a SecurityManager:
- wikimedia/search-repository-swift#20
- wikimedia/search-repository-swift#21

By setting CAN_OVERRIDE_ACCESS_MODIFIERS to false jackson will not try to fix
access modifiers and will work in such environment.