Do not change existing hash password
The original code changes the hash password in the account in case a different password is given than was stored in the account. I can't see why that would be useful as the entire use of a password is then gone. I changed the code to check whether a hash password in the account  already exists. If it does NOT exist and we supply a new hash password, it is set to the given hash password.